add EasifyDefines module and move Constants to EasifyDefines module
