--- Modules/EasifyCore/EasifyCore/Defines/Constants.swift
@@ -1,33 +1,0 @@
-//
-//  Constants.swift
-//  Easify
-//
-//  Created by Muhammed Said Özcan on 17/05/2020.
-//  Copyright © 2020 Muhammed Said Özcan. All rights reserved.
-//
-
-import Foundation
-import SwiftUI
-
-public struct Constants {
-    public struct Files {
-        public static let spotifyPlistFileName: String = "SpotifyCredentials"
-    }
-
-    public struct Copies {
-        public static let connectToSpotifyAccount: String = "Connect to your Spotify account"
-        public static let connectCTA: String = "Connect"
-        public static let easify: String = "Easify"
-        public static let backgroundImageReference: String = "Image: Ron Smith on Unsplash"
-    }
-
-    public struct Spacings {
-        public static let single: CGFloat = 8.0
-        public static let double: CGFloat = 16.0
-        public static let quadruple: CGFloat = 32.0
-    }
-
-    public struct URLS {
-        public static let backgroundImageURL: String = "https://unsplash.com/photos/tknOyEefp2k"
-    }
-}
