+++ Modules/EasifyDefines/EasifyDefines/EasifyDefines.h
@@ -1,0 +1,19 @@
+//
+//  EasifyDefines.h
+//  EasifyDefines
+//
+//  Created by Muhammed Said Özcan on 21/05/2020.
+//  Copyright © 2020 Muhammed Said Özcan. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for EasifyDefines.
+FOUNDATION_EXPORT double EasifyDefinesVersionNumber;
+
+//! Project version string for EasifyDefines.
+FOUNDATION_EXPORT const unsigned char EasifyDefinesVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <EasifyDefines/PublicHeader.h>
+
+
